requirements.txt
astroid==2.3.3
isort==4.3.21
lazy-object-proxy==1.4.3
luma.core==1.12.0
luma.oled==3.3.0
mccabe==0.6.1
pbkdf2==1.3
Pillow==7.0.0
pkg-resources==0.0.0
pyftdi==0.42.2
pylint==2.4.4
pyserial==3.4
pyusb==1.0.2
RPi.GPIO==0.6.5
six==1.14.0
smbus2==0.3.0
spidev==3.4
typed-ast==1.4.1
wifi==0.3.8
wrapt==1.11.2
